+++ ignite/pkg/protoc-gen-dart/data/data_linux_amd64.go
@@ -1,6 +1,0 @@
-package data
-
-import _ "embed" // embed is required for binary embedding.
-
-//go:embed protoc-gen-dart_linux_amd64
-var binary []byte