+++ ignite/services/chain/generate.go
@@ -16,7 +16,6 @@ import (
 
 const (
 	defaultVuexPath    = "vue/src/store"
-	defaultDartPath    = "flutter/lib"
 	defaultOpenAPIPath = "docs/static/openapi.yml"
 )
 
@@ -24,7 +23,6 @@ type generateOptions struct {
 	isGoEnabled       bool
 	isTSClientEnabled bool
 	isVuexEnabled     bool
-	isDartEnabled     bool
 	isOpenAPIEnabled  bool
 	tsClientPath      string
 }
@@ -49,18 +47,11 @@ func GenerateTSClient(path string) GenerateTarget {
 	}
 }
 
-// GenerateTSClient enables generating proto based Typescript Client.
+// GenerateVuex enables generating proto based Typescript Client.
 func GenerateVuex() GenerateTarget {
 	return func(o *generateOptions) {
 		o.isTSClientEnabled = true
 		o.isVuexEnabled = true
-	}
-}
-
-// GenerateDart enables generating Dart client.
-func GenerateDart() GenerateTarget {
-	return func(o *generateOptions) {
-		o.isDartEnabled = true
 	}
 }
 
@@ -88,10 +79,6 @@ func (c *Chain) generateFromConfig(ctx context.Context, cacheStorage cache.Stora
 
 		if conf.Client.Vuex.Path != "" {
 			targets = append(targets, GenerateVuex())
-		}
-
-		if conf.Client.Dart.Path != "" {
-			targets = append(targets, GenerateDart())
 		}
 	}
 
@@ -136,8 +123,8 @@ func (c *Chain) Generate(
 	}
 
 	var (
-		dartPath, openAPIPath, tsClientPath, vuexPath string
-		updateConfig                                  bool
+		openAPIPath, tsClientPath, vuexPath string
+		updateConfig                        bool
 	)
 
 	if targetOptions.isTSClientEnabled {
@@ -187,25 +174,6 @@ func (c *Chain) Generate(
 		)
 	}
 
-	if targetOptions.isDartEnabled {
-		dartPath = conf.Client.Dart.Path
-		if dartPath == "" {
-			dartPath = defaultDartPath
-		}
-
-		dartPath = c.joinGeneratedPath(dartPath)
-		if err := os.MkdirAll(dartPath, 0o766); err != nil {
-			return err
-		}
-
-		options = append(options,
-			cosmosgen.WithDartGeneration(
-				cosmosgen.DartModulePath(dartPath),
-				dartPath,
-			),
-		)
-	}
-
 	if targetOptions.isOpenAPIEnabled {
 		openAPIPath = conf.Client.OpenAPI.Path
 		if openAPIPath == "" {
@@ -243,14 +211,6 @@ func (c *Chain) Generate(
 		if targetOptions.isVuexEnabled {
 			c.ev.Send(
 				fmt.Sprintf("Vuex stores path: %s", vuexPath),
-				events.Icon(icons.Bullet),
-				events.ProgressFinish(),
-			)
-		}
-
-		if targetOptions.isDartEnabled {
-			c.ev.Send(
-				fmt.Sprintf("Dart path: %s", dartPath),
 				events.Icon(icons.Bullet),
 				events.ProgressFinish(),
 			)